# perf(psalm): improve psalm diagnostics performances