--- a/lib/Extension/LanguageServerPsalm/Model/PsalmConfig.php
+++ b/lib/Extension/LanguageServerPsalm/Model/PsalmConfig.php
@@ -4,8 +4,11 @@ namespace Phpactor\Extension\LanguageServerPsalm\Model;
 
 final class PsalmConfig
 {
-    public function __construct(private string $phpstanBin, private bool $shouldShowInfo)
-    {
+    public function __construct(
+        private string $phpstanBin,
+        private bool $shouldShowInfo,
+        private bool $useCache,
+    ) {
     }
 
     public function psalmBin(): string
@@ -16,5 +19,10 @@ final class PsalmConfig
     public function shouldShowInfo(): bool
     {
         return $this->shouldShowInfo;
+    }
+
+    public function useCache(): bool
+    {
+        return $this->useCache;
     }
 }