I've lately moved from [null-ls](https://github.com/jose-elias-alvarez/null-ls.nvim/blob/b287e9d124037a7ea5de8e3368a3b75274cec8b4/lua/null-ls/builtins/diagnostics/psalm.lua) to phpactor for psalm diagnostics and I found it slower...

After investigating, I've found that the `PsalmProcess` class doesn't pass the filename to the command, then, psalm always parse all the project instead of just the file it should analyse.

In my case, It reduces the execution time from 5 to 1.8 seconds.

Then, I've also noticed that `--no-cache` directive is passed to the command, and it's not present in the  [null-ls](https://github.com/jose-elias-alvarez/null-ls.nvim/blob/b287e9d124037a7ea5de8e3368a3b75274cec8b4/lua/null-ls/builtins/diagnostics/psalm.lua) command. I've removed it and it reduces the execution time from 1.8 to 0.9 seconds.