@dantleech 2023-01-03
> Then, I've also noticed that --no-cache directive is passed to the command

This was necessary at the time as the cache didn't invalidate properly. That may have improved.

Possibly not passing the filename was necessary too.

Can you let me know if it works consistently well?