@gbprod 2023-01-04
> This was necessary at the time as the cache didn't invalidate properly. That may have improved.

I don't know if it have been improved, I will try to work without `--no-cache` and look for inconsistencies, but I've used for a long time [null-ls](https://github.com/jose-elias-alvarez/null-ls.nvim/blob/b287e9d124037a7ea5de8e3368a3b75274cec8b4/lua/null-ls/builtins/diagnostics/psalm.lua) and I didn't have any problems...

> Possibly not passing the filename was necessary too.

I don't understand why not passing the filename should be necessary, the results are the same and performances are greatly improved. 
 
I will try to use psalm like this for some times and I'll give you some feedbacks.

PS: I'm running the latest version of psalm.