@dantleech 2023-01-05
> I don't understand why not passing the filename should be necessary

There were lots of issues when I implemented this years ago regarding the cache, it's probably better now.