@gbprod 2023-01-10
> > I don't understand why not passing the filename should be necessary
> 
> There were lots of issues when I implemented this years ago regarding the cache, it's probably better now.

I'm using this branch for 7 days without detecting any issue with psalm diagnostics. 
I don't know if you want to merge like this or if it will be safer to keep the `--no-cache` option ?