@dantleech 2023-01-10
but maybe as a configuration option where the default is "cache enabled" ?