@gbprod 2023-01-16
I've updated the PR to introduce a configuration option to enable/disable cache and it works :)